fix(runner): keep edit retries conflict-driven
